Implemented Zero Tick Bug and Enhanced Ice, Fixed Config Bug

+++ src/main/java/com/klr2003/anaesia/utils/ConfigHandler.java
@@ -12,30 +12,33 @@ public class ConfigHandler {
 
     public static void initConfig() {
 
-        try (OutputStream output = new FileOutputStream(configFileLoc)) {
+        File f = new File(configFileLoc);
+        if(!f.exists()){
+            try (OutputStream output = new FileOutputStream(configFileLoc)) {
 
-            // set the properties value
-            config.put(ConfigList.isSlimeSuperFlatPatchEnabled, "true");
-            config.put(ConfigList.isEnhancedBerriesEnabled, "true");
-            config.put(ConfigList.isEnhancedBoneMealEnabled, "true");
-            config.put(ConfigList.isEnhancedBurningEnabled, "true");
-            config.put(ConfigList.isEnhancedIceEnabled, "true");
-            config.put(ConfigList.isEnhancedKelpEnabled, "true");
-            config.put(ConfigList.isEnhancedSeedsEnabled, "true");
-            config.put(ConfigList.isEnhancedTotemEnabled, "true");
-            config.put(ConfigList.isFishingUnpatchEnabled, "true");
-            config.put(ConfigList.isProtectionUnpatchEnabled, "true");
-            config.put(ConfigList.isWitherRosesUnpatchEnabled, "true");
-            config.put(ConfigList.isZeroTickUnpatchEnabled, "true");
-            config.put(ConfigList.isDebugModeEnabled, "false");
-            config.put(ConfigList.configFileVersion, "1");
+                // set the properties value
+                config.put(ConfigList.isSlimeSuperFlatPatchEnabled, "true");
+                config.put(ConfigList.isEnhancedBerriesEnabled, "true");
+                config.put(ConfigList.isEnhancedBoneMealEnabled, "true");
+                config.put(ConfigList.isEnhancedBurningEnabled, "true");
+                config.put(ConfigList.isEnhancedIceEnabled, "true");
+                config.put(ConfigList.isEnhancedKelpEnabled, "true");
+                config.put(ConfigList.isEnhancedSeedsEnabled, "true");
+                config.put(ConfigList.isEnhancedTotemEnabled, "true");
+                config.put(ConfigList.isFishingUnpatchEnabled, "true");
+                config.put(ConfigList.isProtectionUnpatchEnabled, "true");
+                config.put(ConfigList.isWitherRosesUnpatchEnabled, "true");
+                config.put(ConfigList.isZeroTickUnpatchEnabled, "true");
+                config.put(ConfigList.isDebugModeEnabled, "false");
+                config.put(ConfigList.configFileVersion, "1");
 
-            // save properties to project root folder
-            config.store(output, AnaesiaMod.MOD_NAME + " Mod Config File");
+                // save properties to project root folder
+                config.store(output, AnaesiaMod.MOD_NAME + " Mod Configuration");
 
 
-        } catch (IOException io) {
-            io.printStackTrace();
+            } catch (IOException io) {
+                io.printStackTrace();
+            }
         }
     }
     public static boolean readConfigBoolean(String configString){
